@@ -96,52 +99,56 @@
     "builtin-modules": "^3.3.0",
     "chokidar": "^3.5.3",
     "colorette": "^2.0.20",
-    "concurrently": "^8.1.0",
-    "core-js": "^3.30.2",
+    "concurrently": "^8.2.1",
+    "core-js": "^3.32.2",
     "date-time": "^4.0.0",
     "es5-shim": "^4.6.7",
     "es6-shim": "^0.35.8",
-    "eslint": "^8.42.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-unicorn": "^47.0.0",
-    "eslint-plugin-vue": "^9.14.1",
+    "eslint": "^8.49.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-unicorn": "^48.0.1",
+    "eslint-plugin-vue": "^9.17.0",
     "fixturify": "^3.0.0",
     "flru": "^1.0.2",
     "fs-extra": "^11.1.1",
     "github-api": "^3.4.0",
     "hash.js": "^1.1.7",
     "husky": "^8.0.3",
-    "inquirer": "^9.2.7",
-    "is-reference": "^3.0.1",
-    "lint-staged": "^13.2.2",
-    "locate-character": "^2.0.5",
-    "magic-string": "^0.30.0",
+    "inquirer": "^9.2.11",
+    "is-reference": "^3.0.2",
+    "lint-staged": "^14.0.1",
+    "locate-character": "^3.0.0",
+    "magic-string": "^0.30.3",
     "mocha": "^10.2.0",
     "nyc": "^15.1.0",
-    "pinia": "^2.1.3",
-    "prettier": "^2.8.8",
-    "pretty-bytes": "^6.1.0",
+    "pinia": "^2.1.6",
+    "prettier": "^3.0.3",
+    "pretty-bytes": "^6.1.1",
     "pretty-ms": "^8.0.0",
     "requirejs": "^2.3.6",
-    "rollup": "^3.23.1",
-    "rollup-plugin-license": "^3.0.1",
+    "rollup": "^3.29.2",
+    "rollup-plugin-license": "^3.1.0",
     "rollup-plugin-string": "^3.0.0",
     "rollup-plugin-thatworks": "^1.0.4",
-    "semver": "^7.5.1",
+    "semver": "^7.5.4",
     "shx": "^0.3.4",
-    "signal-exit": "^4.0.2",
+    "signal-exit": "^4.1.0",
     "source-map": "^0.7.4",
     "source-map-support": "^0.5.21",
-    "systemjs": "^6.14.1",
-    "terser": "^5.17.7",
-    "tslib": "^2.5.3",
-    "typescript": "^5.1.3",
-    "vitepress": "^1.0.0-beta.1",
+    "systemjs": "^6.14.2",
+    "terser": "^5.19.4",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9",
+    "vitepress": "^1.0.0-rc.14",
     "vue": "^3.3.4",
     "weak-napi": "^2.0.2",
     "yargs-parser": "^21.1.1"
+  },
+  "overrides": {
+    "semver": "^7.5.4"
   },
   "files": [
     "dist/**/*.js",
