@@ -1,6 +1,6 @@
 {
     "name": "tailwind-merge",
-    "version": "2.2.1",
+    "version": "2.5.2",
     "description": "Merge Tailwind CSS classes without style conflicts",
     "keywords": [
         "tailwindcss",
@@ -55,39 +55,35 @@
         "test:watch": "jest test --config scripts/jest.config.mjs --watch",
         "test:exports": "node scripts/test-built-package-exports.cjs && node scripts/test-built-package-exports.mjs",
         "lint": "eslint --max-warnings 0 '**'",
-        "size": "size-limit",
         "preversion": "if [ -n \"$DANYS_MACHINE\" ]; then git checkout main && git pull; fi",
         "version": "zx scripts/update-readme.mjs",
         "postversion": "if [ -n \"$DANYS_MACHINE\" ]; then git push --follow-tags && open https://github.com/dcastil/tailwind-merge/releases; fi"
     },
-    "dependencies": {
-        "@babel/runtime": "^7.23.7"
-    },
+    "dependencies": {},
     "devDependencies": {
-        "@babel/plugin-transform-runtime": "^7.23.7",
-        "@babel/preset-env": "^7.23.7",
+        "@babel/core": "^7.25.2",
+        "@babel/preset-env": "^7.25.3",
         "@rollup/plugin-babel": "^6.0.4",
         "@rollup/plugin-node-resolve": "^15.2.3",
-        "@rollup/plugin-typescript": "^11.1.5",
-        "@size-limit/preset-small-lib": "^11.0.1",
-        "@types/jest": "^29.5.11",
-        "@typescript-eslint/eslint-plugin": "^6.16.0",
-        "@typescript-eslint/parser": "^6.16.0",
+        "@rollup/plugin-typescript": "^11.1.6",
+        "@types/jest": "^29.5.12",
+        "@typescript-eslint/eslint-plugin": "^8.0.1",
+        "@typescript-eslint/parser": "^8.0.1",
         "babel-plugin-annotate-pure-calls": "^0.4.0",
-        "babel-plugin-polyfill-regenerator": "^0.5.4",
-        "eslint": "^8.56.0",
+        "babel-plugin-polyfill-regenerator": "^0.6.2",
+        "eslint": "^8.57.0",
         "eslint-plugin-import": "^2.29.1",
-        "eslint-plugin-jest": "^27.6.0",
+        "eslint-plugin-jest": "^28.6.0",
         "globby": "^11.1.0",
         "jest": "^29.7.0",
-        "prettier": "^3.1.1",
-        "rollup": "^4.9.2",
+        "prettier": "^3.3.3",
+        "rollup": "^4.19.1",
         "rollup-plugin-delete": "^2.0.0",
-        "rollup-plugin-dts": "^6.1.0",
-        "size-limit": "^11.0.1",
-        "ts-jest": "^29.1.1",
-        "typescript": "^5.3.3",
-        "zx": "^7.2.3"
+        "rollup-plugin-dts": "^6.1.1",
+        "ts-jest": "^29.2.3",
+        "tslib": "^2.6.3",
+        "typescript": "^5.5.4",
+        "zx": "^8.1.4"
     },
     "publishConfig": {
         "provenance": true
